LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/109-expected.txt

Not tested

FAIL  scheduler: javascript URL in iframe, src set via DOM assert_array_equals: property 2, expected "inline script #3" but got "iframe onload"
